requirements.txt
openai-agents==0.*
openai>=1.0.0
PyYAML>=6.0
